- GW attacks closest enemy to the player within a range of 4.5 - Enemy agents now tagged as EnemyAgent instead of Enemy

+++ IAT410/JackHammer/Assets/Scripts/HorsemanAgent.cs
@@ -120,14 +120,13 @@ public class HorsemanAgent : MonoBehaviour
 		}
 	}
 
-	void TakeDamage (int damage)
+	void TakeDamage (float damage)
 	{
         if (state == HorsemanAgent.State.IDLE)
         {
          state = HorsemanAgent.State.CHASE;
         }
         if (health - damage >= 0) {
-   
 			health -= damage;
 			if (gameObject.name != "CannonAgent") {
 				bloodSpawner.SendMessage ("spawn", transform.position, SendMessageOptions.DontRequireReceiver);
